+++ acm/decoder.h
@@ -23,9 +23,9 @@ public:
   void reset();
 
   bool decode(
-    uint8_t *src, unsigned src_len, 
-    uint8_t *dst, unsigned dst_len, 
-    unsigned *src_gone, unsigned *dst_gone);
+    uint8_t *src, size_t src_len, 
+    uint8_t *dst, size_t dst_len, 
+    size_t *src_gone, size_t *dst_gone);
 };
 
 #endif
